feat(pg|tools): add cat-installer for pg (#400)
