+++ .github/workflows/postgres-ded.yml
@@ -7,6 +7,7 @@ on:
 env:
   NODE_VERSION: 14
   PYTHON3_VERSION: 3.8.10
+  VIRTUALENV_VERSION: 20.21.1
 
 jobs:
 
@@ -51,7 +52,7 @@ jobs:
         with:
           python-version: ${{ env.PYTHON3_VERSION }}
       - name: Setup virtualenv
-        run: pip install virtualenv
+        run: pip install virtualenv==${{ env.VIRTUALENV_VERSION }}
       - name: Run deploy
         if: github.event.action == 'synchronize' || github.event.action == 'labeled'
         id: deploy
