+++ .github/workflows/redshift.yml
@@ -18,6 +18,7 @@ env:
   NODE_VERSION: 14
   PYTHON2_VERSION: 2.7.18
   PYTHON3_VERSION: 3.8.10
+  VIRTUALENV_VERSION: 20.21.1
   GCLOUD_VERSION: 290.0.1
 
 jobs:
@@ -52,7 +53,7 @@ jobs:
         with:
           python-version: ${{ env.PYTHON3_VERSION }}
       - name: Setup virtualenv
-        run: pip install virtualenv
+        run: pip install virtualenv==${{ env.VIRTUALENV_VERSION }}
       - name: Configure AWS Credentials
         uses: aws-actions/configure-aws-credentials@v1
         with:
@@ -98,7 +99,7 @@ jobs:
         with:
           python-version: ${{ env.PYTHON3_VERSION }}
       - name: Setup virtualenv
-        run: pip install virtualenv
+        run: pip install virtualenv==${{ env.VIRTUALENV_VERSION }}
       - name: Configure AWS Credentials
         uses: aws-actions/configure-aws-credentials@v1
         with:
@@ -130,7 +131,7 @@ jobs:
         with:
           python-version: ${{ env.PYTHON2_VERSION }}
       - name: Setup virtualenv
-        run: pip install virtualenv
+        run: pip install virtualenv==${{ env.VIRTUALENV_VERSION }}
       - name: Auth google
         uses: google-github-actions/auth@v0
         with:
