+++ tools/installer/.gitignore
