+++ tools/installer/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = cat-installer
-version = 0.1.0
-description = Python script to install the CARTO Analytics Toolbox in Redshift
+version = 0.2.0
+description = Python script to install the CARTO Analytics Toolbox in Redshift and Postgres
 long_description = file: README.md
 long_description_content_type = text/markdown
 keywords =
@@ -32,6 +32,7 @@ install_requires =
     tqdm>4.64
     sqlparse>0.4
     redshift-connector>2.0
+    psycopg2-binary>=2.9.1
 zip_safe = False
 [options.entry_points]
 console_scripts =
