+++ tools/installer/setup.py
